+++ hypertunity/optimisation/bo/_backends/__init__.py
@@ -1,1 +1,0 @@
-from .gpyopt import GPyOptBackend